Add script to export files to RTEMS
This commit adds a script that is a (very hacky) parser for the the
cmake files in the mcux-sdk. It copies all necessary files for RTEMS and
creates yml spec files for the controllers that have been imported.

The script is only intended as a helper for updating the files. It is
not necessary for using the BSP.